requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# hbartax/script.py: 6
python_dotenv == 0.15.0

# hbartax/script.py: 3
requests == 2.24.0
